+++ packages/common/src/GRPCClient.ts
@@ -9,7 +9,8 @@ import * as v1 from './types';
 import * as v2 from '../grpc/v2/concordium/types';
 import { Base58String, HexString, isRpcError } from './types';
 import { QueriesClient } from '../grpc/v2/concordium/service.client';
-import type { RpcTransport } from '@protobuf-ts/runtime-rpc';
+import { HealthClient } from '../grpc/v2/concordium/health.client';
+import type { RpcError, RpcTransport } from '@protobuf-ts/runtime-rpc';
 import { CredentialRegistrationId } from './types/CredentialRegistrationId';
 import * as translate from './GRPCTypeTranslation';
 import { AccountAddress } from './types/accountAddress';
@@ -51,6 +52,7 @@ export type FindInstanceCreationReponse = {
  */
 export class ConcordiumGRPCClient {
     client: QueriesClient;
+    healthClient: HealthClient;
 
     /**
      * Initialize a gRPC client for a specific concordium node.
@@ -58,6 +60,7 @@ export class ConcordiumGRPCClient {
      */
     constructor(transport: RpcTransport) {
         this.client = new QueriesClient(transport);
+        this.healthClient = new HealthClient(transport);
     }
 
     /**
@@ -1502,6 +1505,22 @@ export class ConcordiumGRPCClient {
 
     private async getConsensusHeight() {
         return (await this.getConsensusStatus()).lastFinalizedBlockHeight;
+    }
+
+    /**
+     * Queries the node to check its health
+     *
+     * {@codeblock ~~:client/healthCheck.ts#documentation-snippet}
+     *
+     * @returns a HealthCheck indicating whether the node is healthy or not and provides the message from the client, if not healthy.
+     */
+    async healthCheck(): Promise<v1.HealthCheckResponse> {
+        try {
+            await this.healthClient.check({});
+            return { isHealthy: true };
+        } catch (e) {
+            return { isHealthy: false, message: (e as RpcError).message };
+        }
     }
 }