+++ packages/common/src/types.ts
@@ -2047,3 +2047,12 @@ export type SmartContractTypeValues =
     | number
     | string
     | boolean;
+
+export type HealthCheckResponse =
+    | {
+          isHealthy: true;
+      }
+    | {
+          isHealthy: false;
+          message?: string;
+      };