+++ packages/common/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Added missing fields to `getBlockChainParameters` response. (rootKeys, level1Keys, level2Keys)
+
+### Added
+
+- `sendUpdateInstruction` to the gRPC Client.
+
 ## 9.3.0
 
 ### Added