+++ packages/common/CHANGELOG.md
@@ -3,7 +3,12 @@
 ## Unreleased
 
 ### Added
+
+- `sendUpdateInstruction` to the gRPC Client.
 - `healthCheck` to the gRPC Client.
+
+### Fixed
+- Added missing fields to `getBlockChainParameters` response. (rootKeys, level1Keys, level2Keys)
 
 ## 9.3.0