+++ packages/common/CHANGELOG.md
@@ -6,8 +6,8 @@
 
 - `sendUpdateInstruction` to the gRPC Client.
 - `healthCheck` to the gRPC Client.
--   `Module` class for functionality related to smart contract modules, such as parsing the WebAssembly and interface of the module.
--   Smart contract related types `ContractName`, `EntrypointName` and helper functions `isInitName`, `isReceiveName`, `getContractNameFromInit` and `getNamesFromReceive`.
+- Functions `calculateModuleReference` for getting the module reference and `parseModuleInterface` for getting the interface from the source of a smart contract module.
+- Smart contract related types `ContractName`, `EntrypointName` and helper functions `isInitName`, `isReceiveName`, `getContractNameFromInit` and `getNamesFromReceive`.
 
 ### Fixed
 - Added missing fields to `getBlockChainParameters` response. (rootKeys, level1Keys, level2Keys)